requirements.txt
python-dotenv
requests
user-agents
gunicorn
flask
werkzeug